+++ src/app/cases/page.tsx
@@ -5,7 +5,7 @@ import { UUID } from 'crypto';
 import { CaseListing } from '@/types/schema';
 import { getNCases } from '@/api/supabase/queries/cases';
 import ListingCard from '@/components/ListingCard';
-import CaseDetails from '@/components/CaseDetails';
+import CaseDetails from '@/components/ListingDetails';
 import { H1, H2, CenteredH3 } from '@/styles/text';
 import { ProfileContext } from '@/utils/ProfileProvider';
 import ProfileButton from '@/components/ProfileButton';
