+++ src/components/ListingDetails/styles.tsx
